Update register delegate
